Add a Book to the Cart
This commit introduces the `Order` model to represent the items that a
visitor has added to their cart.

An `Order` record is identifiable by a [secure token], which is
generated `on: :create` by the `has_secure_token` declaration.

As a visitor navigates around the site, they don't yet have an `Order`
record. Once they decide to add a `Book` record to their `Order`, their
action will create a `LineItem` record.

Once the `LineItem` record is persisted, the newly created `Order`
record's `#token` will be persisted to their cookie. Subsequent requests
will include that cookie, which will serve to identify their existing
`Order` record.

The "Add to Cart" button is powered by a [`form_with`][form_with] that
submits to the `line_items#create` controller action. Forms constructed
by [`form_with` calls will default to encoding `data-remote="true"`
attributes][remote-forms], to integrate with [RailsUJS].

These form requests will be made over AJAX. When the
`LineItemsController` responds with a [`redirect_back`][redirect_back],
[Turbolinks will intercede and handle the redirection on the
client-side][turbolinks-redirect] in a way that re-using the existing
page's `<head>` elements.

While the Turbolinks loading might lead to an optimized user experience,
it is helpful to consider the redirect a "full page" reload.

Once a `Book` record is in a visitor's `Order`, the corresponding
`LineItem` record is rendered into the page's cart.

To render the cart, and to handle the expanding and collapsing of the
cart, render a [`<details>`][details] element in the page's `<nav>`,
containing an SVG icon in the child [`<summary>`][summary] element.

To simplify the interactions with the cart in the System Test suite,
introduce the `expand_cart` test helper, which invokes [`find`][find]
with the `text:` option containing the translated `<title>` text we
embed into the cart icon's `inline_svg_tag`.

[secure token]: https://api.rubyonrails.org/v6.0/classes/ActiveRecord/SecureToken/ClassMethods.html#method-i-has_secure_token
[form_with]: https://api.rubyonrails.org/v6.0/classes/ActionView/Helpers/FormHelper.html#method-i-form_with
[remote-forms]: https://guides.rubyonrails.org/v6.0/form_helpers.html#how-do-forms-with-patch-put-or-delete-methods-work-questionmark
[RailsUJS]: https://guides.rubyonrails.org/v6.0/working_with_javascript_in_rails.html#remote-elements
[redirect_back]: https://api.rubyonrails.org/v6.0/classes/ActionController/Redirecting.html#method-i-redirect_back
[turbolinks-redirect]: https://github.com/turbolinks/turbolinks/tree/v5.2.0#redirecting-after-a-form-submission
[details]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details
[summary]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/summary
[find]: https://www.rubydoc.info/github/jnicklas/capybara/master/Capybara/Node/Finders#find-instance_method