[SETUP]: Scaffold `books#index` and `books#show`
As a foundation for the rest of the project, this commit introduces the
skeletal structure for the future application development: a book store.

The pages made available through `/books` and `/books/:id` are
statically declared HTML, served through the `books#index` and
`books#show` view templates, declared in `app/views/books`.

The templates are written in [HTML and ERB][templates]. They are styled
with [CSS] (at this point in the project, the styles are deliberately
_not_ [SCSS]. As an alternative to [SCSS's variable
declarations][scss-variables], the CSS rules make use of [custom
properties] declared on the [`:root` pseudo-selector][root].

While the CSS does not make use of SCSS or other pre-processors, rules
are grouped into separate sub-directories and files, then concatenated
through [Rails' Asset Pipeline][asset-pipeline].

The CSS rules (and their accompanying HTML elements) are written in the
[<abbr title="Blocks, Elements, Modifiers">BEM</abbr> style][BEM].

When text is internationalized through
[`ActionView::Helpers::TranslationHelper#translate`][translate],
configure our `test` and `development` environments [to raise on missing
translations][raise_on_missing_translations].

Finally, to render SVG elements in-line into the HTML response, this
commit adds a dependency on the [`inline_svg` gem][inline_svg].

[templates]: https://guides.rubyonrails.org/v6.0/layouts_and_rendering.html#rendering-by-default-convention-over-configuration-in-action
[CSS]: https://developer.mozilla.org/en-US/docs/Web/CSS
[SCSS]: https://sass-lang.com/guide
[scss-variables]: https://sass-lang.com/guide#topic-2
[custom properties]: https://developer.mozilla.org/en-US/docs/Web/CSS/--*
[root]: https://developer.mozilla.org/en-US/docs/Web/CSS/:root
[asset-pipeline]: https://guides.rubyonrails.org/v6.0/asset_pipeline.html#how-to-use-the-asset-pipeline
[BEM]: http://getbem.com/introduction/
[translate]: https://api.rubyonrails.org/v6.0/classes/ActionView/Helpers/TranslationHelper.html#method-i-translate
[raise_on_missing_translations]: https://thoughtbot.com/blog/foolproof-i18n-setup-in-rails#raise-an-exception-on-missing-translations
[inline_svg]: https://github.com/jamesmartin/inline_svg